feat(bookmark modal): Add bookmark editing functionality
This commit adds functionality to edit bookmarks and their associated tags. The modal now allows the user to update the title, url and tags for a bookmark. A set of icons have been added to manage the tags, allowing the user to add, edit and delete tags associated with the bookmark.